Fix publish-baseline preflight: read schema_migrations.version as i32 (int4)
bootstrap_preflight read coalesce(max(version),0) into i64, but
schema_migrations.version is integer (int4); rust-postgres is strict on
column type, so .get::<_,i64>(0) panicked ('error deserializing column 0')
and publish-baseline died at the first preflight query against the live DB.
Bind as i32 (matching the migration applier + CURRENT_SCHEMA_VERSION).

Audited every other new bootstrap query against the real CT110 schema
(package_catalog columns, count(*) int8, nullable digests, jsonb): no other
type mismatches. Verified read-only via pg_typeof/information_schema.

Surfaced because the DB-backed loopback/gated tests require pgvector/pg_search
and skip in the dev env; validated here directly against the live schema.

+++ crates/jurisearch-package-build/src/cycle.rs
@@ -818,14 +818,16 @@ fn bootstrap_preflight(
 ) -> Result<(), BuildError> {
     let mut db = producer.client()?;
 
-    // 1. The producer schema must be exactly what this binary builds for.
-    let schema_version: i64 = db
+    // 1. The producer schema must be exactly what this binary builds for. `schema_migrations.version`
+    //    is `integer` (int4), so it MUST be read as `i32` — reading it as `i64` panics in rust-postgres
+    //    with a deserialization error (the migration applier reads it the same way).
+    let schema_version: i32 = db
         .query_one(
             "SELECT coalesce(max(version), 0) FROM schema_migrations;",
             &[],
         )?
         .get(0);
-    let expected = i64::from(jurisearch_storage::migrations::CURRENT_SCHEMA_VERSION);
+    let expected = jurisearch_storage::migrations::CURRENT_SCHEMA_VERSION;
     if schema_version != expected {
         return Err(bootstrap_fail(format!(
             "producer schema_version {schema_version} != CURRENT_SCHEMA_VERSION {expected}; migrate \
